Move game logic in environment and entities
The game_handler module is removed. The environment is now in charge of
calling entities update method and entities are in charge of applying
their effects on the environment
It is now possible to plant a bomb on top of another bomb. We need to
think about what is the best behaviour