Modifications to config loader and ui loader
Config submodules do not need to add themselves to the list of config
submodules in config_loader.
The exception is the UI config submodules, since they are not always
needed and may crash the program.
The config module does not import them automatically anymore for the
same reasons.
Same for the ui module
The package entry point now checks for ImportError
The package setup now demands a python version >= 3.9
The package setup tool is repared (disfunctional since 0.5.0)